task2-preprocessing/requirements.txt
pandas==2.1.4
numpy==1.26.4
paho-mqtt==1.6.1
pika==1.3.2
matplotlib==3.7.2
plotly==5.18.0
prophet==1.1.5
scikit-learn==1.3.0

